+++ tests/Rule/CountTest.php
@@ -4,13 +4,14 @@ declare(strict_types=1);
 
 namespace Yiisoft\Validator\Tests\Rule;
 
+use ArrayIterator;
 use Countable;
 use stdClass;
 use Yiisoft\Validator\DataSet\SingleValueDataSet;
 use Yiisoft\Validator\Rule\Count;
 use Yiisoft\Validator\Rule\CountHandler;
 use Yiisoft\Validator\Tests\Rule\Base\DifferentRuleInHandlerTestTrait;
-use Yiisoft\Validator\Tests\Rule\Base\LimitTestTrait;
+use Yiisoft\Validator\Tests\Rule\Base\CountableLimitTestTrait;
 use Yiisoft\Validator\Tests\Rule\Base\RuleTestCase;
 use Yiisoft\Validator\Tests\Rule\Base\RuleWithOptionsTestTrait;
 use Yiisoft\Validator\Tests\Rule\Base\SkipOnErrorTestTrait;
@@ -19,8 +20,8 @@ use Yiisoft\Validator\Tests\Support\Data\CountDto;
 
 final class CountTest extends RuleTestCase
 {
+    use CountableLimitTestTrait;
     use DifferentRuleInHandlerTestTrait;
-    use LimitTestTrait;
     use RuleWithOptionsTestTrait;
     use SkipOnErrorTestTrait;
     use WhenTestTrait;
@@ -71,10 +72,14 @@ final class CountTest extends RuleTestCase
         return [
             [[0, 0, 0], [new Count(min: 3)]],
             [[0, 0, 0, 0], [new Count(min: 3)]],
-            [[0, 0, 0], [new Count(exactly: 3)]],
+            [[0, 0, 0], [new Count(3)]],
             [[], [new Count(max: 3)]],
             [[0, 0], [new Count(max: 3)]],
             [[0, 0, 0], [new Count(max: 3)]],
+            'value: array iterator with min allowed count, min: positive' => [
+                new ArrayIterator([0, 0, 0]),
+                [new Count(min: 3)],
+            ],
             [
                 new SingleValueDataSet(
                     new class () implements Countable {
@@ -90,6 +95,26 @@ final class CountTest extends RuleTestCase
             ],
             'class attribute' => [
                 new CountDto(7),
+            ],
+            'value: empty array, exactly: 0' => [
+                [],
+                [new Count(0)],
+            ],
+            'value: empty array, min: 0' => [
+                [],
+                [new Count(min: 0)],
+            ],
+            'value: empty array, max: 0' => [
+                [],
+                [new Count(max: 0)],
+            ],
+            'value: empty array iterator, exactly: 0' => [
+                new ArrayIterator(),
+                [new Count(0)],
+            ],
+            'value: empty array iterator, exactly: positive, skipOnEmpty: true' => [
+                new ArrayIterator(),
+                [new Count(1, skipOnEmpty: true)],
             ],
         ];
     }
@@ -129,8 +154,13 @@ final class CountTest extends RuleTestCase
             [[''], [new Count(min: 3)], ['' => [$lessThanMinmessage]]],
             [['some string'], [new Count(min: 3)], ['' => [$lessThanMinmessage]]],
             [[new stdClass()], [new Count(min: 3)], ['' => [$lessThanMinmessage]]],
+            'value: array iterator with lower count, min: positive' => [
+                new ArrayIterator([0, 0]),
+                [new Count(min: 3)],
+                ['' => [$lessThanMinmessage]],
+            ],
             // https://www.php.net/manual/ru/class.countable.php
-            [
+            'value: class with min count returned from count method but not implenting Countable interface, min: 3' => [
                 [
                     new class () {
                         protected int $myCount = 3;
@@ -190,7 +220,7 @@ final class CountTest extends RuleTestCase
 
             'custom not exactly message' => [
                 [0, 0, 0, 0],
-                [new Count(exactly: 3, notExactlyMessage: 'Custom not exactly message.')],
+                [new Count(3, notExactlyMessage: 'Custom not exactly message.')],
                 ['' => ['Custom not exactly message.']],
             ],
             'custom not exactly message with parameters' => [
@@ -217,6 +247,16 @@ final class CountTest extends RuleTestCase
                 new CountDto(),
                 null,
                 ['' => ['This value must contain at least 2 items.']],
+            ],
+            'value: array with greater count, exactly: 0' => [
+                [0],
+                [new Count(0)],
+                ['' => ['This value must contain exactly 0 items.']],
+            ],
+            'value: empty array iterator, exactly: positive' => [
+                new ArrayIterator(),
+                [new Count(1)],
+                ['' => ['This value must contain exactly 1 item.']],
             ],
         ];
     }