+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.0.1 under development
 
 - Enh #567: Add immutable setter for modifying `Validator::$defaultSkipOnEmptyCondition` (@arogachev)
+- Bug #586: Allow `0` as a limit in `CountableLimitTrait` (@arogachev)
+- Bug #431: Add support for empty iterables to `WhenEmpty` (@arogachev)
 
 ## 1.0.0 February 22, 2023